Revert "Projects Page (#268)"
This reverts commit f3a19accbc828e6330a004694cf27db0f29e75d2.
