+++ apps/web/package.json
@@ -25,7 +25,6 @@
     "dotenv": "^16.4.5",
     "geist": "^1.3.0",
     "lucide-react": "0.363.0",
-    "mini-svg-data-uri": "^1.4.4",
     "negotiator": "0.6.3",
     "next": "14.1.1",
     "next-auth": "5.0.0-beta.16",
