+++ apps/web/src/app/[lang]/layout.tsx
@@ -23,7 +23,7 @@ export default async function RootLayout({
   if (!resources) return <></>;
   return (
     <html>
-      <body className={`${GeistSans.className}`} data-app-name={appName}>
+      <body className={`${GeistSans.className} overflow-hidden`}>
         <Providers lang={params.lang}>{children}</Providers>
       </body>
     </html>
