+++ apps/web/src/app/[lang]/public/page.tsx
@@ -1,11 +1,11 @@
-"use server";
+import { Brands } from "@repo/ui/brands";
+import { Hero } from "@repo/ui/hero";
 
-import { isCustomNodeSelected } from "@tiptap-location/lib/utils";
-import Landing from "./landing";
-import { getConfig } from "./layout";
-
-export default async function Page() {
-  let appName = process.env?.APPLICATION_NAME || "konya";
-  let config = await getConfig(appName);
-  return <Landing config={config} />;
+export default function Page(): JSX.Element {
+  return (
+    <div className="w-full bg-red-200 p-0 h-[200vh]">
+      <Hero />
+      <Brands />
+    </div>
+  );
 }
