+++ apps/web/src/app/globals.css
@@ -50,13 +50,3 @@
 * {
   scroll-behavior: smooth;
 }
-
-.no-visible-scrollbar {
-  scrollbar-width: none;
-  -ms-overflow-style: none;
-  -webkit-overflow-scrolling: touch;
-}
-
-.no-visible-scrollbar::-webkit-scrollbar {
-  display: none;
-}
