+++ packages/ui/package.json
@@ -3,8 +3,7 @@
   "version": "0.0.0",
   "exports": {
     "./test": "./src/test.tsx",
-    "./*": "./src/*/index.tsx",
-    "./upwithcrowd/*": "./src/upwithcrowd/*/index.tsx"
+    "./*": "./src/*/index.tsx"
   },
   "license": "MIT",
   "scripts": {
@@ -14,7 +13,6 @@
     "type-check": "tsc --noEmit"
   },
   "peerDependencies": {
-    "@ayasofyazilim/saas": "^0.0.13",
     "@repo/ayasofyazilim-ui": "workspace:*",
     "next": "14.1.1",
     "react": "18.2.0",
@@ -39,9 +37,6 @@
     "@dnd-kit/sortable": "^8.0.0",
     "@dnd-kit/utilities": "^3.2.2",
     "embla-carousel-auto-scroll": "^8.1.3",
-    "framer-motion": "^11.2.10",
-    "lucide-react": "^0.379.0",
-    "mini-svg-data-uri": "^1.4.4",
-    "simplex-noise": "^4.0.1"
+    "lucide-react": "^0.379.0"
   }
 }
