+++ packages/ui/src/language-selector/index.tsx
@@ -23,7 +23,7 @@ export default function LanguageSelector({
       <TooltipTrigger asChild>
         <div className="flex">
           <CountrySelector
-            searchText={resources?.AbpUi?.texts?.Search || "Search"}
+            searchText={resources?.AbpUi.texts?.Search}
             searchEmptyValue={
               resources?.AbpExceptionHandling?.texts?.DefaultErrorMessage404
             }
